index-update

+++ MainApp/views.py
@@ -1,8 +1,10 @@
 from django.http import Http404
+from django.contrib.auth.decorators import login_required
+from django.core.exceptions import PermissionDenied
 from django.shortcuts import render, redirect
 from django.contrib import auth
 from MainApp.models import Snippet
-from MainApp.forms import SnippetForm, UserRegistrationForm
+from MainApp.forms import SnippetForm, UserRegistrationForm, CommentForm
 
 
 def index_page(request):
@@ -31,22 +33,36 @@ def snippets_page(request):
     snippets = Snippet.objects.all()
     context = {
         'pagename': 'Просмотр сниппетов',
-        "snippets": snippets
+        'snippets': snippets,
+    }
+    return render(request, 'pages/view_snippets.html', context)
+
+
+@login_required
+def snippets_my(request):
+    my_snippets = Snippet.objects.filter(user=request.user)
+    context = {
+        'pagename': 'Мои сниппеты',
+        "snippets": my_snippets
     }
     return render(request, 'pages/view_snippets.html', context)
 
 
 def snippet_detail(request, snippet_id):
     snippet = Snippet.objects.get(id=snippet_id)
+    comment_form = CommentForm()
     context = {
         'pagename': 'Информация о сниппете',
-        "snippet": snippet
+        "snippet": snippet,
+        "comment_form": comment_form
     }
     return render(request, 'pages/snippet-detail.html', context)
 
 
 def snippet_delete(requests, snippet_id):
     snippet = Snippet.objects.get(id=snippet_id)
+    if snippet.user != requests.user:
+        raise PermissionDenied()
     snippet.delete()
     return redirect('snippets-list')
 
@@ -81,10 +97,24 @@ def registration(request):
         form = UserRegistrationForm(request.POST)
         if form.is_valid():
             form.save()
-        else:
+        else:  # данные не валидные
             context = {
                 'pagename': 'Регистрация',
                 "form": form
             }
             return render(request, 'pages/registration.html', context)
         return redirect('home')
+
+
+def comment_create(request):
+    if request.method == "POST":
+        comment_form = CommentForm(request.POST)
+        snippet_id = request.POST.get("snippet_id")
+        if comment_form.is_valid():
+            comment = comment_form.save(commit=False)
+            comment.author = request.user
+            snippet = Snippet.objects.get(id=snippet_id)
+            comment.snippet = snippet
+            comment.save()
+        raise
+        return redirect(request.META.get('HTTP_REFERER', '/'))
